+++ environments/environment.ts
@@ -2,5 +2,14 @@ export const environment = {
     production: true,
     storageURL:'https://storage.googleapis.com',
     bucketName:'dev-sg-dashboard',
-    folderName:'sg-dashboard'
+    folderName:'sg-dashboard',
+    firebaseConfig:{
+        apiKey: "API_KEY",
+        authDomain: "AUTH_DOMAIN",
+        projectId: "PROJECT_ID",
+        storageBucket: "STORAGE_BUCKET",
+        messagingSenderId: "MESSAGE_SENDER_ID",
+        appId: "APP_ID",
+        measurementId: "MEASUREMENT_ID"
+    }
 };
