+++ src/app/app.config.ts
@@ -5,16 +5,9 @@ import { provideAnimations } from '@angular/platform-browser/animations';
 import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
 import { getFirestore, provideFirestore } from '@angular/fire/firestore';
 import { routes } from './app.routes';
+import { environment } from '../../environments/environment';
 
-const firebaseConfig = {
-  apiKey: "API_KEY",
-  authDomain: "AUTH_DOMAIN",
-  projectId: "PROJECT_ID",
-  storageBucket: "STORAGE_BUCKET",
-  messagingSenderId: "MESSAGE_SENDER_ID",
-  appId: "APP_ID",
-  measurementId: "MEASUREMENT_ID"
-};
+const firebaseConfig = environment.firebaseConfig
 
 export const appConfig: ApplicationConfig = {
   providers: [
