+++ src/app/components/stories-carousel/stories-carousel.component.html
@@ -13,7 +13,7 @@
              [ngClass]="{'active': i === currentChunkIndex}">
 
           <div class="story-card" *ngFor="let slide of chunk">
-            <app-improvement-story [story]="slide"></app-improvement-story>
+            <app-improvement-story [story]="slide" [browserId]="browserId"></app-improvement-story>
           </div>
         </div>
       </div>
